platform: tegra: mc: Fix Cert-C error handling bug
In function mc_translate_update_ch casting mc_generate_chan_offset()
from unsigned long long to unsigned int without checking its value may
result in lost or misinterpreted data. Fix this by changing data type
of variable.

CID 370548
Bug 3512545

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2738884
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/platform/tegra/mc/mc_addr_translate.c
@@ -192,7 +192,7 @@ static u64 mc_translate_update_ch(u64 addr, u32 ch)
 	u32 cur_chan;
 	u32 chanbits = 2;
 	u32 chanpos = 10;
-	u32 chanoffset;
+	u64 chanoffset;
 
 	for (cur_chan = 0; (cur_chan < (1 << chanbits)); ++cur_chan) {
